refactor: types 파일 제거
- IError는 공통이므로 common 파일로 분리
- 불필요해진 type 제거하면서 types 파일 제거함
