--- src/api/types.ts
@@ -1,9 +1,0 @@
-export interface IError {
-  statusCode: number;
-  data: unknown | null;
-}
-
-export interface IImagesPreSignedUrlResponse {
-  url: string;
-  key: string;
-}
