+++ src/api/types/common.ts
@@ -1,0 +1,4 @@
+export interface IError {
+  statusCode: number;
+  data: unknown | null;
+}
